Let Riak assign a random key
This change makes it possible to pass None as the key argument
to bucket.new(). I'm not sure if you'll want to change the
request method to POST, but it seemed like the simplest way
to make this work. An alternative would be to change the request
method to POST whenever the key is None, or to refactor the HTTP
transport so that a post method is called instead.