Pass a ConnectionManager to the transport constructor.
Create an appropriate CM and pass that to the transport. We also adjust
the constructor signature and allow for arbitrary options to be passed to
the transport (eg. Http takes a prefix and mapred_prefix). Some backwards
compat code has been introduced in all transports (other than the basic
RiakHttpTransport) to extract a host/port pair from the CM.

This commit also re-enables testing of RiakHttpReuseTransport, even though
it is busted and should go away "soon".