Make the protobuf transport use the ConnectionManager
This uses a subclass of the generic connection.Socket since we need to
record the connection-stateful value for client_id. There are
certainly issues around using variant client_id values in a pooled
environment (as noted by the old CachedTransport, and documented
within the set_client_id method here).

The send/recv sections now use cm.withconn() to acquire and use a
connection, which is then passed down into the methods which do the
actual send/recv of the bits.

This also switches over to sock.sendall() to try and avoid short writes.

The host/port has been stripped out, since that is now managed within
the connection manager.