Fixes issue #124
This is what I believe the best behaviour. The get(None) doesn't throw
an error, to be consistent.

However I'm not sure about delete(). To be completely consistent,
.delete() shouldn't throw any errors either, as the key of None doesn't
exists in the database anyway. However, the Pbc client currently throws
a TypeError (actually I would like it better if it didn't throw any
error and just moved on). Any input?