Add pagination to secondary indexes, commit the first.
This introduces the IndexPage class, which is necessary both to wrap
the results into a sensible API, and to provide consistency between
single-roundtrip and streaming options. The IndexPage class can appear
as if it were a list, is iterable for streaming and regular iteration
purposes. Regardless of whether pagination is used, an IndexPage will
be returned (thanks to the quirkiness of Python's generators).

This leaves equality/return-terms dangling, to be addressed in the
next commit. Sending an equality query with the return-terms flag on
does not result in pairs being returned, but simply keys. Obviously,
the client knows what the index term is and so can inject it into the
result (which is what we will do for consistency).