Add pagination to secondary indexes, commit the second.
This refactors a bit of the IndexPage class so as to support the
equality/return-terms issue left dangling in the last commit. Along
the way, I added a test for streaming with the eq/rt combo, and
discovered an issue where the PBC interface returned the `keys` field
from the response directly when it was expected to be a list. This
wouldn't normally be an issue, but was necessitated by the eq/rt
issue.

This completes the work for #252.